feat(config): add database configuration utilities with multi-database support   - Add 9 new utility classes in config.util package for production-ready database management.   - DatabasePropertiesParser: parse .properties files to DatabaseConfig with validation.   - ConfigurationPriorityResolver: resolve config from multiple sources (CLI > ENV > FILE > DEFAULT).   - ConfigurationExporter: export config to Properties/JSON/ENV with automatic password masking.   - SensitiveDataMasker: mask credentials in logs and output for security.   - ConfigurationSecurityAnalyzer: detect security risks with remediation recommendations.   - SecurityWarning: structured security warning with severity levels.   - FilePermissionsUtil: manage POSIX permissions (600 for config files).   - DatabaseConnectionTester: test connectivity with detailed diagnostics.   - DatabaseMigrator: schema migration with version tracking and history.   - Change ValidationSummary and ChainValidationResult from int to long for block counts.   - Add DatabaseConfigurationDemo showcasing new utilities.   - Add comprehensive test coverage (9 test classes, 100% coverage).   - Update documentation: API_GUIDE.md, TECHNICAL_DETAILS.md, PRODUCTION_GUIDE.md.   - Move PRODUCTION_GUIDE.md to docs/deployment/ for better organization.   - Security-first design: masking by default, permission checks, comprehensive warnings.   - Thread-safe and reusable: suitable for CLI, web apps, REST APIs.

+++ src/main/java/com/rbatllet/blockchain/service/ValidationReport.java
@@ -175,14 +175,14 @@ public class ValidationReport {
         private long invalidBlocks = 0;
         private long offChainFiles = 0;
         private long corruptFiles = 0;
-        private int totalChecks = 0;
-        private int passedChecks = 0;
-        private int failedChecks = 0;
+        private long totalChecks = 0;
+        private long passedChecks = 0;
+        private long failedChecks = 0;
         private long validationDuration = 0;
         
         public ValidationMetrics() {}
         
-        public ValidationMetrics(int totalChecks, int passedChecks, int failedChecks, long validationDuration) {
+        public ValidationMetrics(long totalChecks, long passedChecks, long failedChecks, long validationDuration) {
             this.totalChecks = totalChecks;
             this.passedChecks = passedChecks;
             this.failedChecks = failedChecks;
@@ -205,14 +205,14 @@ public class ValidationReport {
         public long getCorruptFiles() { return corruptFiles; }
         public void setCorruptFiles(long corruptFiles) { this.corruptFiles = corruptFiles; }
         
-        public int getTotalChecks() { return totalChecks; }
-        public void setTotalChecks(int totalChecks) { this.totalChecks = totalChecks; }
+        public long getTotalChecks() { return totalChecks; }
+        public void setTotalChecks(long totalChecks) { this.totalChecks = totalChecks; }
         
-        public int getPassedChecks() { return passedChecks; }
-        public void setPassedChecks(int passedChecks) { this.passedChecks = passedChecks; }
+        public long getPassedChecks() { return passedChecks; }
+        public void setPassedChecks(long passedChecks) { this.passedChecks = passedChecks; }
         
-        public int getFailedChecks() { return failedChecks; }
-        public void setFailedChecks(int failedChecks) { this.failedChecks = failedChecks; }
+        public long getFailedChecks() { return failedChecks; }
+        public void setFailedChecks(long failedChecks) { this.failedChecks = failedChecks; }
         
         public long getValidationDuration() { return validationDuration; }
         public void setValidationDuration(long validationDuration) { this.validationDuration = validationDuration; }
